Move `search_for_adt_without_structural_match` to `ty/mod`
